Add build ROCM support to build-triton-wheel.yml
To match with upstream and build triton whl's locally so nightly pytorch whls can
access them without needing to use pypi.org.

We may have a better approach to build both whl's at once, but for now, to save
duplication of code, another matrix is added for device (cuda/rocm)  With rocm
invoking a different commit and repo.  The goal is to eventually have a single
whl support both backends.